+++ .prettierrc.js
@@ -1,0 +1,13 @@
+/********************************\
+ * Prettier configuration file. *
+\********************************/
+
+/** @type {import("@types/prettier").RequiredOptions} */
+module.exports = {
+  arrowParens: "always",
+  bracketSpacing: true,
+  printWidth: 80,
+  semi: true,
+  singleQuote: false,
+  trailingComma: "none"
+}