+++ postcss.config.js
@@ -1,0 +1,11 @@
+/*******************************\
+ * PostCSS configuration file. *
+\*******************************/
+
+/** @type {import("postcss").Processor} */
+module.exports = {
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
+}